make DHW entity a fixture
